22/7/25

+++ src/main/java/me/ajg/diss/encoders/dnaFountain/DNAFountainEncoder.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public class DNAFountainEncoder {
     
-    public static List<String> encode(String filePath, int INPUT_SEED, double redundancy, int fragmentLength, float maxGcDeviation, int maxHomopolymer){
+    public static List<String> encode(String filePath, int INPUT_SEED, double redundancy, int fragmentLength, double maxGcDeviation, int maxHomopolymer){
         //Compress to gzip
         byte[] gzipCompressed = FileUtil.gzipCompression(FileUtil.fileToByteArray(filePath));
         //fragment the data
